+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.1.0](https://github.com/in2workspace/in2-verifier-api/releases/tag/v1.1.0)
+### Added
+- Add refresh token support for the OpenID Connect flow
+- Add nonce support for the OpenID Connect authorization code flow
+
 ## [v1.0.16](https://github.com/in2workspace/in2-verifier-api/releases/tag/v1.0.16)
 ### Fixed
 - Add time window validation for the credential in the Verifiable Presentation